abdallah
create cart and cartItems and order and orderItems services

+++ backEnd/library/src/main/java/com/library_management_system/library/entity/OrderItem.java
@@ -27,8 +27,8 @@ public class OrderItem {
     private Book product;
 
     private int quantity;
-    private BigDecimal price;
-    private BigDecimal fine;
+    private double price;
+    private double fine;
 
     // Getters and setters
 }
